README.md: Add a brief description for the purpose of the repo
The purpose of this repo is to share image_sign.sh and the associated CSF
file standalone.

The various examples of signing code on the i.MX devices rely on hard-coded
addresses in the CSF file.

With a little bit of extra work though the hard-coding of addresses can be
rightly thrown over-board.

Ahoy !

Signed-off-by: Bryan O'Donoghue <pure.logic@nexus-software.ie>