bl2_sign.csf: Add a CSF file with a HAB_BLOCKS_REPLACE token
The token HAB_BLOCKS_REPLACE is used to allow image_sign.sh to replace the
token with HAB Blocks: 0xNumber 0xNumber 0xNumber as the NXP CST tool
requires.

This token means we don't need to hard-code anything in the CSF file since
doing so is just asking for trouble and continuous repetitive updates of
the CSF file for no purpose at all.

Signed-off-by: Bryan O'Donoghue <pure.logic@nexus-software.ie>